work around working-dir bug in structure/biotac
Both these drivers use the current directory during initialization to
search for libraries. The flow code screws this up by changing directory
into the episode dir. The workaround is to cache the original directory
when the program starts, and temporarily change back to it during
initialization of these drivers.

These drivers must NEVER be initialized concurrently!! I should likely
put in a mutex to prevent this.

This is definitely a Schroedinbug. It's totally impossible that the
stickcam and optocam episodes worked until I discovered the problem
while testing biocam.